requirements.txt
requests
urllib3
qrcode
pywebview
mutagen